pyproject.toml
[tool.poetry]
name = "receipt_processor_service"
version = "0.1.0"
description = ""
authors = ["zacharyrsmith99 <zachary.r.smith99@gmail.com>"]
readme = "README.md"
packages = [{include = "receipt_processor_service"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.114.1"}
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "receipt_processor_service.main:main"
